requirements.txt
boto3==1.21.22
botocore==1.24.22
greenlet==1.1.2
jmespath==1.0.0
packaging==21.3
Pillow==9.0.1
psycopg2==2.9.3
pure-python-adb==0.3.0.dev0
pyparsing==3.0.7
pytesseract==0.3.9
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.32
urllib3==1.26.9
